test-requirements.txt
pytest
mock
requests-mock
pandas